+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * `ls` now supports options  `-p/--prefix` and `-a/--start-after` to list only files starting with a prefix or start listing from a specified key
 * new sub-command `acl` to get and set ACL's
 
+
 ## 0.4.0 ⛸
 * New sub-command `get` with option `-h` to return HeadObject.
 * New sub-command `share` to create presigned URL's.