+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 0.6.0 🛺
+* Added sub-comand `mb` (Make a bucket)
+
 ## 0.5.0 🛵
 * removed ending `/` when listing buckets
 * fixed sub-command `get` to download file in the current path